Added Basic functionality of Deposit Pit, Coins, Gacha and Cosmetics.
Created a deposit pit that players can through either junk, sellables, or themselves into. Created a new InteractableRigidbody class that will be on all grabbable rigidbodies, this holds the value of the item. Coins are spat out from the deposit hole. You can take a coin to the Gacha machine to recieve a random cosmetic. Cosmetics are currently just shapes that the player can wear on their heads.